+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 All notable changes to this project will be documented in this file.
 This project *tries* to adhere to [Semantic Versioning](http://semver.org/), even before v1.0.
 
-## [UNRELEASED]
+## [1.2.0]
+- #7 from patchkit-net/feature/validations
 - Add a way to quickly define getters and setters using `column` method
 - Can be used with `ActiveModel::Validations`
 - Add ActiveModel::Validations (group=test,development) dependency